app/backend/scripts/requirements.txt
requests
python-dotenv
pandas
tqdm
google-generativeai
Flask


